app/upload_to_zenodo/requirements.txt
pynodo==0.2.2
pyyaml==6.0
